+++ src/main/java/br/edu/ifpb/es/daw/entities/Turma.java
@@ -6,22 +6,22 @@ import jakarta.persistence.*;
 public class Turma {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String nome;
     private Integer ano;
 
 
-    public Turma(int id, String nome, Integer ano) {
+    public Turma(Long id, String nome, Integer ano) {
         this.id = id;
         this.nome = nome;
         this.ano = ano;
     }
 
-    public int getId() { return id; }
+    public Long getId() { return id; }
     public String getNome() { return nome; }
     public Integer getAno() { return ano; }
 
-    public void setId(int id) { this.id = id; }
+    public void setId(Long id) { this.id = id; }
     public void setNome(String nome) { this.nome = nome; }
     public void setAno(Integer ano) { this.ano = ano; }
